VSL-98: Moved some parts of the code to utils, removed some styling that was not necessary
