+++ packages/client/src/utils.js
@@ -72,9 +72,8 @@ export const getProgressPercentageForSignersAmount = (signersAmount) => {
 };
 
 export const getFlowscanUrlForTransaction = (hash) => {
-  return `https://${
-    process.env.REACT_APP_FLOW_ENV === "mainnet" ? "" : "testnet."
-  }flowscan.org/transaction/${hash}`;
+  return `https://${process.env.REACT_APP_FLOW_ENV === "mainnet" ? "" : "testnet."
+    }flowscan.org/transaction/${hash}`;
 };
 
 export const syncSafeOwnersWithSigners = (signers, safeOwners) => {
@@ -100,3 +99,18 @@ export const getVaultId = (identifiers, coinType) => {
 };
 
 export const removeAddressPrefix = (address) => address.replace("0x", "");
+
+export const getTokenMeta = (vaultId) => {
+  if (vaultId) {
+    const tokenName = vaultId.split(".")[2];
+    return COIN_TYPE_TO_META[CONTRACT_NAME_TO_COIN_TYPE[tokenName]];
+  }
+}
+
+export const getNFTMeta = (collectionId) => {
+  const NFTName = collectionId?.split(".")[2];
+  const NFTAddress = collectionId?.split(".")[1];
+  return {
+    NFTName, NFTAddress
+  };
+}
